Moves the annotation to the good package.
